Split post 1426 oversized slices
